Move utility function tests over to new .py file
